third_party/WebKit/LayoutTests/fast/html/imports/window-onload-with-import-stylesheet.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
window.jsTestIsAsync = true;

window.onload = function(event)
{
    shouldBeTrue("target.import.querySelector('link').sheet instanceof StyleSheet");
    finishJSTest();   
}
</script>
<link id="target" rel="import" href="resources/hello-css.html">
